Update and rename Baekjoon/boj2306/main.cpp to Baekjoon/boj2306/mingyu/boj2306.cpp
